RLIB-16: Better fix for the previous one.
When UTF-8 is enabled, r_strlen() is g_utf8_strlen() that returns
a signed value. When UTF-8 is disabled, r_strlen() is plain strlen()
and it returns an unsigned value. Cast the left side of the comparison
to signed and modify the type of "n" to ssize_t to keep it signed
and make it change its size just like size_t and glong.

+++ libsrc/pcode_op_functions.c
@@ -1476,9 +1476,9 @@ gint rlib_pcode_operator_left(rlib *r, struct rlib_pcode *code, struct rlib_valu
 
 	if(RLIB_VALUE_IS_STRING(v2) && RLIB_VALUE_IS_NUMBER(v1) && RLIB_VALUE_GET_AS_STRING(v2) != NULL) {
 		gchar *tmp = g_strdup(RLIB_VALUE_GET_AS_STRING(v2));
-		gint n = RLIB_VALUE_GET_AS_NUMBER(v1)/RLIB_DECIMAL_PRECISION;
+		ssize_t n = RLIB_VALUE_GET_AS_NUMBER(v1)/RLIB_DECIMAL_PRECISION;
 		if (n >= 0) {
-			if (r_strlen(tmp) > (guint)n) *r_ptrfromindex(tmp, n) = '\0';
+			if ((ssize_t)r_strlen(tmp) > n) *r_ptrfromindex(tmp, n) = '\0';
 		}
 		rlib_value_free(v1);
 		rlib_value_free(v2);
